+++ src/Component.js
@@ -1,0 +1,26 @@
+import {createElement} from "./utils";
+
+export class Component {
+  constructor() {
+    if (new.target === Component) {
+      throw new Error(`Can't instantiate Component, only concrete one.`);
+    }
+  }
+
+  get template() {
+    throw new Error(`You have to define template.`);
+  }
+
+  bind() {}
+
+  set onClick(func) {
+    this._onClickHandler = func;
+  }
+
+  render() {
+    this._element = createElement(this.template).children[0];
+    this.bind();
+    return this._element;
+  }
+}
+
