+++ src/card/Card.js
@@ -1,8 +1,9 @@
 import {createCardTemplate} from './create-template';
-import {createElement} from "../utils";
+import {Component} from "../Component";
 
-export class Card {
+export class Card extends Component {
   constructor(data) {
+    super();
     this._data = data;
     this._element = null;
     this.onClickHandler = this.onClickHandler.bind(this);
@@ -17,21 +18,11 @@ export class Card {
     this._onClickHandler(this.getData());
   }
 
-  set onClick(func) {
-    this._onClickHandler = func;
-  }
-
   bind() {
     this._element.querySelector(`.film-card__comments`).addEventListener(`click`, this.onClickHandler);
   }
 
   getData() {
     return Object.assign({}, this._data);
-  }
-
-  render() {
-    this._element = createElement(this.template).children[0];
-    this.bind();
-    return this._element;
   }
 }
