+++ src/popup-card/Popup-Card.js
@@ -1,8 +1,9 @@
 import {createPopupTemplate} from './create-popup-template';
-import {createElement} from "../utils";
+import {Component} from "../Component";
 
-export class Popup {
+export class Popup extends Component {
   constructor(data) {
+    super();
     this._data = data;
     this._element = null;
     this._closeBtn = null;
@@ -13,10 +14,6 @@ export class Popup {
     evt.preventDefault();
     this.unbind();
     this._onClickHandler(this._element);
-  }
-
-  set onClick(func) {
-    this._onClickHandler = func;
   }
 
   bind() {
@@ -31,11 +28,5 @@ export class Popup {
 
   get template() {
     return createPopupTemplate(this._data);
-  }
-
-  render() {
-    this._element = createElement(this.template).children[0];
-    this.bind();
-    return this._element;
   }
 }
